+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * New lint cop `SpaceBeforeFirstArg` checks for space between the method name and the first argument in method calls without parentheses. ([@jonas054][])
 * New style cop `SingleSpaceBeforeFirstArg` checks that no more than one space is used between the method name and the first argument in method calls without parentheses. ([@jonas054][])
 * New formatter `disabled_lines` displays cops and line ranges disabled by inline comments. ([@fshowalter][])
+* New cop `UselessAccessModifiers` checks for access modifiers that have no effect. ([@fshowalter][])
 
 ### Changes